migrate response processing to QtiCreator's plugins
