+++ views/js/qtiCreator/plugins/panel/tpl/dialog.tpl
