+++ views/js/qtiCreator/plugins/panel/tpl/trigger.tpl
